### 2019-01-27 #40 @whedon
[PRE REVIEW]: A course on Geographic Data Science
**Submitting author:** @darribas (<a href="http://orcid.org/0000-0002-6274-1619">Daniel Arribas-Bel</a>)
**Repository:** <a href="https://github.com/darribas/gds18" target ="_blank">https://github.com/darribas/gds18</a>
**Version:** v4.0
