### 2019-01-27 #40 @labarba
👋  @lheagy — Would you like to review this submission to JOSE? Looks right up your alley! (If you're unable, could you recommend someone else? Thanks!)
