### 2019-01-27 #40 @lheagy
Hi @labarba: This looks really interesting! I would be happy to review :)
