### 2019-01-28 1cac024
Adding ISBN/DOI to references
