+++ paper.bib
@@ -3,7 +3,8 @@
       title={Designing Learning Objects for Online Learning},
         author={Norman, Solvig and Porter, David},
           year={2007},
-            journal={Commonwealth of Learning}
+            journal={Commonwealth of Learning},
+      url={http://hdl.handle.net/11599/45 }
 }
 
 @article{hutchings2006principles,
@@ -11,7 +12,8 @@
         author={Hutchings, B},
           journal={Centre for Excellence in Enquiry-Based Learning Resources,
               University of Manchester, England},
-            year={2006}
+            year={2006},
+            url={www.ceebl.manchester.ac.uk/resources/papers/ceeblgr002.pdf}
 }
 
 @article{lage2000inverting,
@@ -22,7 +24,8 @@
   number={1},
   pages={30--43},
   year={2000},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.2307/1183338}
 }
 
 @article{donoho201750,
@@ -33,14 +36,16 @@
   number={4},
   pages={745--766},
   year={2017},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/10618600.2017.1384734}
 }
 
 @book{kitchin2014data,
   title={The data revolution: Big data, open data, data infrastructures and their consequences},
   author={Kitchin, Rob},
   year={2014},
-  publisher={Sage}
+  publisher={Sage},
+  doi={ISBN:9781446287484}
 }
 
 @article{anselin1999interactive,
@@ -50,7 +55,8 @@
   volume={1},
   pages={251--264},
   year={1999},
-  publisher={Citeseer}
+  publisher={Citeseer},
+  doi={ISBN:978-0-471-73545-8}
 }
 
 @article{duque2007supervised,
@@ -61,14 +67,16 @@
   number={3},
   pages={195--220},
   year={2007},
-  publisher={Sage Publications Sage CA: Los Angeles, CA}
+  publisher={Sage Publications Sage CA: Los Angeles, CA},
+  doi={10.1177/0160017607301605 }
 }
 
 @book{webber2018predictive,
   title={The Predictive Postcode: The Geodemographic Classification of British Society},
   author={Webber, Richard and Burrows, Roger},
   year={2018},
-  publisher={SAGE}
+  publisher={SAGE},
+  doi={ISBN:9781526402349},
 }
 
 @book{harris2005geodemographics,
@@ -76,7 +84,8 @@
   author={Harris, Richard and Sleight, Peter and Webber, Richard},
   volume={7},
   year={2005},
-  publisher={John Wiley and Sons}
+  publisher={John Wiley and Sons},
+  doi={ISBN: 978-0-470-86413-5}
 }
 
 @book{boots1988point,
@@ -84,5 +93,6 @@
   author={Boots, Barry N and Getis, Arthur},
   volume={8},
   year={1988},
-  publisher={Sage Publications, Inc}
+  publisher={Sage Publications, Inc},
+  doi={ISBN:0803922450}
 }
