### 2019-01-28 #40 @labarba
@darribas : You shouldn't use the `doi` field for ISBNs (that turns into a broken link). The two "missing DOIs" suggested by `whedon` are false leads.
