### 2019-01-28 #40 @labarba
Also, you need { } around GIS in https://github.com/darribas/gds18/blob/1cac024f860caa9bbd31b47df937b9377daf724a/paper.bib#L83
