### 2019-01-28 #40 @jsta
I'm happy to contribute a review on this paper @labarba
