### 2019-01-28 #42 @whedon
[REVIEW]: A course on Geographic Data Science
**Submitting author:** @darribas (<a href="http://orcid.org/0000-0002-6274-1619">Daniel Arribas-Bel</a>)
**Repository:** <a href="https://github.com/darribas/gds18" target ="_blank">https://github.com/darribas/gds18</a>
**Version:** v4.1
