### 2019-01-30 darribas/gds18#1 @lheagy
JOSE review: Licensing 
This is related to: openjournals/jose-reviews#42

The license specified in the [LICENSE](https://github.com/darribas/gds18/blob/master/LICENSE) file (CC-BY-NC-SA) is not the same as that listed in the [README](https://github.com/darribas/gds18/blob/master/README.md#license) (CC-BY), so they should be updated to be consistent.

With respect to license choice, I would encourage you to consider not using the `NC` portion of the Creative Commons License, as it is fairly restrictive on how others can use the content you developed (see the creative commons explanation on "Free Cultural Work": https://creativecommons.org/share-your-work/public-domain/freeworks)
