### 2019-02-06 darribas/gds18#2 @jsta
Include `dill` and `rtree` in conda file?
This is related to: [openjournals/jose-reviews#42](https://github.com/openjournals/jose-reviews/issues/42)

I get a warning in several notebooks about not having `dill` installed. Even if not needed it might be good to include so learners don't think there is a problem with their installation.

Lab 3 mentions using `rtree` for spatial joins.
