### 2019-02-06 darribas/gds18#3 @jsta
Set default data folder to `data`?
This is related to: [openjournals/jose-reviews#42](https://github.com/openjournals/jose-reviews/issues/42)

I see that several of the notebooks use a default data path of something like `../../../gds17_data`. It's great that there is a walkthrough for learners having difficulty with their data path but my feeling is that using `data` as the default would be preferable. The repo would then work "out of the box".
