### 2019-02-10 darribas/gds18#6 @lheagy
installation instructions
I am following the Linux / OS instructions. This is minor, but for activating a conda environment, the preferred way is now `conda activate` rather than `source activate` (as of [conda 4.4.0](https://github.com/conda/conda/blob/master/CHANGELOG.md#new-feature-highlights-2)). Other than that everything looks good :)

(part of the JOSE review in openjournals/jose-reviews#42)
