### 2019-02-10 darribas/gds18#7 @lheagy
suggestion: downloading data in the labs
I started working through Lab_01, and it took me a bit to find the correct web-page and link to follow to download the csv file for use in the lab. There are a couple options for simplifying this,

1. provide the url in the notebook with instructions on where to put the file so that it can be read-in

2. programatically download the data. This can be done with `requests` and has the added benefit that it simplifies reproducibility of the workflow (which is a nice meta-lesson in best practices)
```python
import requests     # This allows us to download files

f = 'liv_pop.csv'                            # Path to file containing the table
url = "http://darribas.org/gds18/content/labs/data/liv_pop.csv"

# if the file doesn't already exsist, then download it
if not os.path.exists(f):
    request = requests.get(url)
    with open(f, 'wb') as file:
        file.write(request.content)

db = pd.read_csv(f, index_col='GeographyCode')    # Read the table in

```

(related to the JOSE review in openjournals/jose-reviews#42)
