### 2019-02-10 darribas/gds18#8 @lheagy
Lab 3: no module pysal.contrib
I looks like the API for pysal may have changed:
```
<ipython-input-1-1b56deea36c3> in <module>
      4 import geopandas as gpd
      5 import pysal as ps
----> 6 from pysal.contrib.viz import mapping as maps
      7 import palettable as pltt
      8 from seaborn import palplot

ModuleNotFoundError: No module named 'pysal.contrib'
```

This is a minor issues related to the JOSE review: openjournals/jose-reviews#42 (not a publication blocker)
