### 2019-02-10 darribas/gds18#9 @lheagy
Community guidelines
In the repo, I did not see community guidelines (e.g. a Community.md or a Contributing.md) that provide guidelines for others interested in contributing (e.g. fixing typos, reporting bugs, ...). The JOSE requirement is

**Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support

related to the review in openjournals/jose-reviews#42
