### 2019-02-10 darribas/gds18#10 @lheagy
tagged version of the repository
Although the version number is stated as v4.0 in JOSE review (openjournals/jose-reviews#42), there is no indication of this in the github repository. A release needs to be created in order for the repository to have a version number (and it doesn't need to start with 1, it is totally fine for your first release to be v4.0

![image](https://user-images.githubusercontent.com/6361812/52541179-ef769f00-2d46-11e9-88b3-ffbaba192dd9.png)
