### 2019-02-10 #42 @lheagy
Congrats on a really nice course @darribas! They are interesting examples and are well-explained. I have a few review comments, and once these are addressed I think it will be ready for publication

## review comments
- [ ] darribas/gds18#1: the repo license (cc-by-nc-sa) is different than the license stated in the README and on the website (cc-by) (@labarba, does JOSE have any concerns with the 'NC' portion of the license as it is [not considered "Free Cultural Work"](https://creativecommons.org/share-your-work/public-domain/freeworks)?)
- [ ] darribas/gds18#10: there is no version information in the github repository
- [ ] darribas/gds18#9: community guidelines. I did not see these anywhere in the repo
- [ ] darribas/gds18#7: a suggestion on simplifying the process for downloading data in the labs. Right now, the learner has to follow at least 2 links to get to the data from the notebook in most cases; this overhead could be reduced so there is less startup time for the labs.

### minor fixes (not publication blockers in my opinion)
- [ ] darribas/gds18#6: installation instructions `conda activate` is preferred over `source activate`
- [ ] darribas/gds18#8: it looks like the api of pysal changed, so a couple of the labs error on import

## feedback on the review process
- the paper criteria `Does the paper tell the "story" of how the authors came to develop it, or what their expertise is?` is a bit vague to me. I quite like the paper heading "Experience of use" used by @darribas and think that this captures the spirit of this criteria. So it might be worth considering that language as a bold-faced title for this bullet-point
