### 2019-02-12 #42 @jsta
This paper and course material looks quote good to me. I focused my review on the `labs` portion of the repository. In addition to the comments by @lheagy, I had several ideas for improvement:

# review comments

* Does the conda file specify all the dependencies including ones required for the optional excercises? (https://github.com/darribas/gds18/issues/2)

# minor fixes

* In the lab01 notebook, dropping columns is shown using both `del foo` and `pandas.drop` syntax. Consider only using panadas syntax to decrease cognative load?

* The paper states that the target audience is learners with little to no prior knowledge. Maybe text to this effect could be added to the `Overview` page?

* There are multiple instances where an operation is claimed to be simple or easy. I would recommend not emphasizing this so much as it can de-motivate people who are struggling.

* The default data path is highly variable among notebooks. Maybe these could be made consistent? (https://github.com/darribas/gds18/issues/3)
