### 2019-02-28 eda7b3e
Updating license in slides
