### 2019-02-28 e5b5417
Updating license in remaining bits and pieces
