+++ _data/license.yml
@@ -1,3 +1,3 @@
 # Licensing information for course content
-name: Creative Commons BY 4.0
-url: http://creativecommons.org/licenses/by/4.0/
+name: Creative Commons BY-SA 4.0
+url: https://creativecommons.org/licenses/by-sa/4.0/
