### 2019-02-28 darribas/gds18#1 @darribas
Thank you very much for catching this and suggesting a solution! I've decided to go for CC-BY-SA.

In the process of fixing this, I've also realised licensing was very inconsistent across the content. Hopefully e5b54178a3d9feed6d9059f56dff0c0e66265c88, eda7b3e80dd23f7cbd26b29dd511d4c7b9e27349 and 3b5f9569270edb026cf68b85732f68e639271e82 address this.

[Closing but feel free to re-open if anything is unclear or missing]
