### 2019-02-28 darribas/gds18#10 @darribas
Excellent catch! I've added a release with `v4.0` tag:

> https://github.com/darribas/gds18/releases/tag/v4.0

Once (if) the paper is accepted, I'll push `v.4.1` if there's been more changes to reflect those. Leaving open until that happens.
