### 2019-02-28 23b5eab
Pinning to PySAL 1.14 to ensure imports work. Addressing #8
