+++ content/infrastructure/install_gds_stack.yml
@@ -31,7 +31,7 @@ dependencies:
       - geopy
       - polyline
   - pyhamcrest
-  - pysal
+  - pysal=1.14
   - pytest
   - qgrid
   - rasterio
