### 2019-02-28 darribas/gds18#8 @darribas
Good catch! Since the course finished this year (December'18) and now, [we released](https://twitter.com/darribas/status/1090913357509607424) PySAL 2.0 which introduces a whole bunch of API changes. Since this iteration of the course was designed and delivered on PySAL 1.X, I've decided to leave it like that and pin to that version on the installer. Next year's iteration will be based on PySAL 2.X API.

[Closing for now but feel free to re-open if needed]
