### 2019-02-28 17a0644
Replacing source activate by conda activate. Addresses #6
