### 2019-02-28 darribas/gds18#6 @darribas
Great catch! Addressed in #17a064450a9793c778640c5a04e3044fd714e612

[Closing for now but feel free to re-open if need be]
