### 2019-02-28 1e358ff
Removing mention to rtree as it gets automatically installed. Addresses part of #2
