### 2019-02-28 darribas/gds18#2 @darribas
Thanks very much for the comment!

* On `dill`, can you provide a specific example where that happens? I have not run into that, from memory. I'm happy to dig further.
* On `rtree`, it actually now installs automatically when `geopandas` is installed via conda:

```shell
(gds) dh073194:gds18 dani$ conda list | grep rtree
rtree                     0.8.3                 py36_1000    conda-forge
(gds) dh073194:gds18 dani$
```
Also, it is a library the end-user does not need to interact with. For those reasons, I've removed the mention in Lab 3 (1e358ff43bf3590baa3d593d4e8c94553c9a8776).
