### 2019-02-28 2157538
Adding reference to no prior knowledge required in the Overview page. This addresses Minor Fix.2 by @jsta in #42
