### 2019-02-28 d4d8e52
Adding a bit more explanation to make clear that del and drop are not the same and have their own purpose. This addresses minor fix.1 by @jsta
