### 2019-02-28 #42 @darribas
@jsta: on your minor fix.1, a bit of context: `del foo` and `pandas.drop` do not the same thing exactly. The former removes "in-place" while the latter creates a copy without the columns dropped. Both are useful in different contexts. Based on that, I've decided not to change it. However, I added a bit more of text to illustrate the differences and make them come out more clearly ([d4d8e522f14fce5ebe09f3f5bced20ce3f9eb289](https://github.com/darribas/gds18/commit/d4d8e522f14fce5ebe09f3f5bced20ce3f9eb289)).
