### 2019-02-28 darribas/gds18#2 @jsta
Hmm, I can no longer reproduce the `dill` warnings. Looks good on `rtree` :+1:
