### 2019-03-01 5858998
Replacing simple, easy, trivial by more accurate terms in lab 9
