### 2019-03-01 #42 @darribas
@jsta: on your minor fix.3, I've just gone over every notebook and [replaced/removed](https://github.com/darribas/gds18/compare/d4d8e522f14fce5ebe09f3f5bced20ce3f9eb289...585899888fa13329606e11a5bd4a3f6bc1ec9f24) any instance of "simple", "easy", and "trivial". This has been an incredible wake-up to how the language can influence the student's attitude. I think it's a lot better now. Thanks!
