### 2019-03-01 efecd9c
Adding CONTRIBUTING guidelines borrowed from the rcourse. This addresses #9 by @lheagy
