### 2019-03-01 darribas/gds18#9 @darribas
Great comment! I've now added a `CONTRIBUTING.md` file borrowed and adapted from the [`UofTCoders/rcourse`](https://github.com/UofTCoders/rcourse).

Closing for now but feel free to reopen if needed.
