### 2019-03-01 7a76786
Adding link to path tutorial in Lab 2. Part of the fix for #3 and #7
