### 2019-03-01 484f14a
Adding web links to path tutorial in Lab 1. Part of the fix for #3 and #7
