### 2019-03-01 4bd9389
Adding web links to path tutorial in Lab 6. Part of the fix for #3 and #7
