### 2019-03-01 78f0c71
Adding web links to path tutorial in Lab 1 (recompiling HTML+PDF). Part of the fix for #3 and #7
