### 2019-03-01 dcd507e
Adding web links to path tutorial in Lab 8. Part of the fix for #3 and #7
