### 2019-03-01 4f1945b
Adding web links to path tutorial in Lab 6. Part of the fix for #3 and #7
