### 2019-03-01 #42 @darribas
Hello @lheagy and @jsta, I have looked at your issues about the links and access to the data ([`#3`](https://github.com/darribas/gds18/issues/3) and [`#7`](https://github.com/darribas/gds18/issues/7). Very good points both. A few thoughts and things I have(n't) done:

- There is one case (ie. Twitter dataset in Lab 9) where the license of the data does not allow public repackaging. In those cases, I'm hosting it on a private site for students only.
- There are several cases where the datasets come from a portal ([CDRC data](https://data.cdrc.ac.uk/)) that requires a free login to download datasets. I would like keep these cases as they are because:
     a) Those logins are used for statistics reported back to the funding body of the portal so I think it is important to give credit instead of downloading once and linking to the Github repo (effectively repackaging data).
     b) I think it is also important from the pedagogical point of view that the student gets some experience in accessing "real-world" data in the way they would have to if they were looking for data outside a structured course
- In cases where the data is already repackaged (ie. in the `data` folder), I have updated the notebooks to illustrate how, in addition to reading the file locally if you have downloaded it, you can read it from the course website link. This is implemented in commits from [`7a767`](https://github.com/darribas/gds18/commit/7a7678606b09275fbd26d59fc396d7e0ba5c7381) to [`4f194`](https://github.com/darribas/gds18/commit/4f1945b71b58636e9133247c1f8d506ebb28c6ec).
- In response to @lheagy's suggestion to write a bespoke method to download it in the notebook, I've decided against it, although I think it can be a useful approach in other contexts. My main motivation for this is:
     a) The pedagogical argument mentioned above of "forcing" students to get out in the wild to access data
     b) My sense is the code, even if explicitly mentioned as optional, would feel too advanced and put off students somehow
