### 2019-03-01 darribas/gds18#3 @darribas
I've responded to this in [this comment](https://github.com/openjournals/jose-reviews/issues/42#issuecomment-468757751).
