### 2019-03-01 darribas/gds18#7 @darribas
I've responded to this [here](https://github.com/openjournals/jose-reviews/issues/42#issuecomment-468757751). Closing now but feel free to re-open.
