### 2019-03-01 #42 @darribas
@labarba, @lheagy, @jsta. THANK you very much for going over my paper and project. I really appreciate your insight and comments. I think I have addressed every comment. Where the reviewer had opened an issue, I've responded there; where there was a comment in this thread, I've responded accordingly too. I think that should cover all of your concerns but I'm more than happy to work further on aspects you consider need it. I'll wait for your response, and thank you very much again!
