### 2019-03-02 darribas/gds18#7 @lheagy
Thanks for taking time to respond - I totally appreciate your point about learning to do things "in the real world"!
