### 2019-03-22 #42 @lheagy
Congrats on a great course @darribas! All good on my end @labarba 🎉
