### 2019-03-22 #42 @jsta
I checked the remaining items in my review list. I think the paper is good to go. :shipit:
