### 2019-03-26 #42 @labarba
¶1
conscutive >> typo: consecutive

¶2
Although the content presented is relevant… >> the content presented where? The point you're trying to make here is unclear. And perhaps you could start this paragraph with the second sentence. (Or, you might just delete the first sentence.)
