### 2019-03-26 #42 @labarba
https://github.com/darribas/gds18/pull/11
