### 2019-03-26 #42 @labarba
"Although the learning emphasis is placed first and foremost on the student, there is support provided mainly through two channels: within the lab, there is at least one instructor for every 15 students answering questions one-on-one or in small groups; beyond contact time, the class has access to an online discussion forum monitored by the course leader, where students are encouraged to post not only questions but also answers and responses both to other students and to their original enquiry, in case they solve it themselves."

This is one LONG sentence. (1) can you cut this up? (2) can you remove syntactic expletives? (there is, there are); (3) can you simplify? E.g., "the learning emphasis is placed first and foremost on the student"—is this hyperbole to say "student-centered"?
