### 2019-03-26 #42 @labarba
@darribas — You have some bad DOIs in your BibTex. Please check/fix.
