### 2019-04-01 bbc71df
Fixing identified by @lbarba on #42#issuecomment-476791410 for conscutive
