### 2019-04-01 ff476de
Changing h. for h when referring to hour, following suggestion by @labarba in #13
