### 2019-04-01 f7e15f3
Addressing long sentence picked up by @labarba in #42#issuecomment-476800591
