### 2019-04-01 05ed87e
Attempt to fix DOIs picked up by @whedon on review
