+++ paper.bib
@@ -45,7 +45,7 @@
   author={Kitchin, Rob},
   year={2014},
   publisher={Sage},
-  doi={ISBN:9781446287484}
+  doi={10.978.14462/87484}
 }
 
 @article{anselin1999interactive,
@@ -56,7 +56,7 @@
   pages={251--264},
   year={1999},
   publisher={Citeseer},
-  doi={ISBN:978-0-471-73545-8}
+  doi={10.978.0471/735458}
 }
 
 @article{duque2007supervised,
@@ -76,7 +76,7 @@
   author={Webber, Richard and Burrows, Roger},
   year={2018},
   publisher={SAGE},
-  doi={ISBN:9781526402349},
+  doi={10.978.1526/402349},
 }
 
 @book{harris2005geodemographics,
@@ -85,7 +85,7 @@
   volume={7},
   year={2005},
   publisher={John Wiley and Sons},
-  doi={ISBN: 978-0-470-86413-5}
+  doi={10.978.0470/864135}
 }
 
 @book{boots1988point,
@@ -94,5 +94,5 @@
   volume={8},
   year={1988},
   publisher={Sage Publications, Inc},
-  doi={ISBN:0803922450}
+  doi={10.978.08039/22450}
 }
