### 2019-04-01 #42 @darribas
Hello @labarba, thank you very much for the review! I've gone over your comments, and I think I've included them all. The only one I'm unsure is the one relating the DOIs. The ones that come as invalid are all books, so they don't technically have a DOI but a ISBN. Following [this](https://www.doi.org/factsheets/ISBN-A.html), I've tried to turn them into actionable ISBNs (ISBN-A), but not sure if with full success. If there's any guidance in the journal about how to deal with these, I'll be happy to follow it if what I've done doesn't work.
