### 2019-04-03 #42 @labarba
These links are not resolving:

https://doi.org/10.978.0471/735458
https://doi.org/10.978.08039/22450
https://doi.org/10.978.0470/864135
https://doi.org/10.978.14462/87484

If the reference doesn't have a DOI, then don't provide a DOI.
