### 2019-04-07 8a83b98
Removing failing DOIs
