+++ paper.bib
@@ -44,8 +44,7 @@
   title={The data revolution: Big data, open data, data infrastructures and their consequences},
   author={Kitchin, Rob},
   year={2014},
-  publisher={Sage},
-  doi={10.978.14462/87484}
+  publisher={Sage}
 }
 
 @article{anselin1999interactive,
@@ -55,8 +54,7 @@
   volume={1},
   pages={251--264},
   year={1999},
-  publisher={Citeseer},
-  doi={10.978.0471/735458}
+  publisher={Citeseer}
 }
 
 @article{duque2007supervised,
@@ -84,8 +82,7 @@
   author={Harris, Richard and Sleight, Peter and Webber, Richard},
   volume={7},
   year={2005},
-  publisher={John Wiley and Sons},
-  doi={10.978.0470/864135}
+  publisher={John Wiley and Sons}
 }
 
 @book{boots1988point,
@@ -93,6 +90,5 @@
   author={Boots, Barry N and Getis, Arthur},
   volume={8},
   year={1988},
-  publisher={Sage Publications, Inc},
-  doi={10.978.08039/22450}
+  publisher={Sage Publications, Inc}
 }
