### 2019-04-07 #42 @darribas
Thanks very much @labarba. I've just removed the failing DOIs with [8a83b98](https://github.com/darribas/gds18/commit/8a83b9814f8eaaf4c13133aa08fc95622ee97983). Let me know if there's anything else I should do.
