### 2019-04-20 #42 @labarba
@darribas — Before we publish your JOSE paper, we'll need: 1) a new tagged release; please report the version number here, and 2) an archival deposit in Zenodo, Figshare or a similar service; please report the archive DOI here.
