### 2019-04-25 a693a6a
Adding DOI to README
