### 2019-04-25 #42 @labarba
Can you edit the metadata of the Zenodo deposit so the title and author list matches the paper? Thanks!
