+++ paper.bib
@@ -13,7 +13,7 @@
           journal={Centre for Excellence in Enquiry-Based Learning Resources,
               University of Manchester, England},
             year={2006},
-            url={www.ceebl.manchester.ac.uk/resources/papers/ceeblgr002.pdf}
+            url={http://www.ceebl.manchester.ac.uk/resources/papers/ceeblgr002.pdf}
 }
 
 @article{lage2000inverting,
@@ -78,7 +78,7 @@
 }
 
 @book{harris2005geodemographics,
-  title={Geodemographics, GIS and neighbourhood targeting},
+  title={Geodemographics, {GIS} and neighbourhood targeting},
   author={Harris, Richard and Sleight, Peter and Webber, Richard},
   volume={7},
   year={2005},
