### 2019-04-25 #42 @labarba
Would you accept these little tweaks to your bib?

https://github.com/darribas/gds18/pull/14
